fix compilation with new new nightly VaList
In 2025-12-05, a Rust pull request at
https://github.com/rust-lang/rust/pull/141980
reworked the `VaList` implementation to be abi-compatible with C. This
broke:

  * the `printf-compat` implementation, which is fixed in version 0.3
  * pretty much every function that methods of `VaList` (most commonly
    `as_va_list`), since they no longer exist

Here, we fix both of those problems and update the compiler version in
`rust-toolchain.toml` to be the oldest compiler that will already have
this new implementation.

Updating the compiler has created 2 unforseen problems:

  1. When unwinding, the new Rust version now prints the process's PID.
     This means for the `c-gull-unwinding` we can no longer use the
     `assert_cmd` API, as it does not expose the child PID to us.
  2. The `c-gull/src/nss.rs` had 3 extra warnings of operations whose
     results were unused. This was also breaking some other tests.

Both of these were also fixed.